refactor: move UpdateLicenses into setup.py
